requirements.txt
torch
torchvision
numpy
opencv-python
pillow
matplotlib
pyzmq
requests
furhat-remote-api
